Reshape Parkings API
