+++ internal/parkings/data.go
@@ -1,0 +1,85 @@
+package parkings
+
+import (
+	"fmt"
+	"strconv"
+	"time"
+)
+
+// Parking defines details and spaces available for P+R parkings
+type Parking struct {
+	ID                        string    `json:"Id"`
+	Label                     string    `json:"label"`
+	UpdatedTime               time.Time `json:"updated_time"`
+	AvailableStandardSpaces   int       `json:"available_space"`
+	AvailableAccessibleSpaces int       `json:"available_accessible_space"`
+	TotalStandardSpaces       int       `json:"available_normal_space"`
+	TotalAccessibleSpaces     int       `json:"total_space"`
+}
+
+type ByParkingId []Parking
+
+func (p ByParkingId) Len() int           { return len(p) }
+func (p ByParkingId) Less(i, j int) bool { return p[i].ID < p[j].ID }
+func (p ByParkingId) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
+
+// NewParking creates a new Parking object based on a line read from a CSV
+func NewParking(record []string, location *time.Location) (*Parking, error) {
+	if len(record) < 8 {
+		return nil, fmt.Errorf("Missing field in Parking record")
+	}
+
+	updatedTime, err := time.ParseInLocation("2006-01-02 15:04:05", record[2], location)
+	if err != nil {
+		return nil, err
+	}
+	availableStd, err := strconv.Atoi(record[4])
+	if err != nil {
+		return nil, err
+	}
+	totalStd, err := strconv.Atoi(record[5])
+	if err != nil {
+		return nil, err
+	}
+	availableAcc, err := strconv.Atoi(record[6])
+	if err != nil {
+		return nil, err
+	}
+	totalAcc, err := strconv.Atoi(record[7])
+	if err != nil {
+		return nil, err
+	}
+
+	return &Parking{
+		ID:                        record[0],    // COD_PAR_REL
+		Label:                     record[1],    // LIB_PAR_REL
+		UpdatedTime:               updatedTime,  // DATEHEURE_COMPTAGE
+		AvailableStandardSpaces:   availableStd, // NB_TOT_PLACE_DISPO
+		AvailableAccessibleSpaces: availableAcc, // NB_TOT_PLACE_PMR_DISPO
+		TotalStandardSpaces:       totalStd,     // CAP_VEH_NOR
+		TotalAccessibleSpaces:     totalAcc,     // CAP_VEH_PMR
+	}, nil
+}
+
+// ParkingLineConsumer constructs a parking from a slice of strings
+type ParkingLineConsumer struct {
+	parkings map[string]Parking
+}
+
+func makeParkingLineConsumer() *ParkingLineConsumer {
+	return &ParkingLineConsumer{
+		parkings: make(map[string]Parking),
+	}
+}
+
+func (p *ParkingLineConsumer) Consume(line []string, loc *time.Location) error {
+	parking, err := NewParking(line, loc)
+	if err != nil {
+		return err
+	}
+
+	p.parkings[parking.ID] = *parking
+	return nil
+}
+
+func (p *ParkingLineConsumer) Terminate() {}
