+++ type.go
@@ -11,6 +11,7 @@ import (
 	"github.com/CanalTP/forseti/internal/departures"
 	"github.com/CanalTP/forseti/internal/equipments"
 	"github.com/CanalTP/forseti/internal/freefloatings"
+	"github.com/CanalTP/forseti/internal/parkings"
 )
 
 var spFileName = "mapping_stops.csv"
@@ -22,84 +23,6 @@ type VehicleOccupancyRequestParameter struct {
 	VehicleJourneyId string
 	Date             time.Time
 }
-
-// Parking defines details and spaces available for P+R parkings
-type Parking struct {
-	ID                        string    `json:"Id"`
-	Label                     string    `json:"label"`
-	UpdatedTime               time.Time `json:"updated_time"`
-	AvailableStandardSpaces   int       `json:"available_space"`
-	AvailableAccessibleSpaces int       `json:"available_accessible_space"`
-	TotalStandardSpaces       int       `json:"available_normal_space"`
-	TotalAccessibleSpaces     int       `json:"total_space"`
-}
-
-type ByParkingId []Parking
-
-func (p ByParkingId) Len() int           { return len(p) }
-func (p ByParkingId) Less(i, j int) bool { return p[i].ID < p[j].ID }
-func (p ByParkingId) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
-
-// NewParking creates a new Parking object based on a line read from a CSV
-func NewParking(record []string, location *time.Location) (*Parking, error) {
-	if len(record) < 8 {
-		return nil, fmt.Errorf("Missing field in Parking record")
-	}
-
-	updatedTime, err := time.ParseInLocation("2006-01-02 15:04:05", record[2], location)
-	if err != nil {
-		return nil, err
-	}
-	availableStd, err := strconv.Atoi(record[4])
-	if err != nil {
-		return nil, err
-	}
-	totalStd, err := strconv.Atoi(record[5])
-	if err != nil {
-		return nil, err
-	}
-	availableAcc, err := strconv.Atoi(record[6])
-	if err != nil {
-		return nil, err
-	}
-	totalAcc, err := strconv.Atoi(record[7])
-	if err != nil {
-		return nil, err
-	}
-
-	return &Parking{
-		ID:                        record[0],    // COD_PAR_REL
-		Label:                     record[1],    // LIB_PAR_REL
-		UpdatedTime:               updatedTime,  // DATEHEURE_COMPTAGE
-		AvailableStandardSpaces:   availableStd, // NB_TOT_PLACE_DISPO
-		AvailableAccessibleSpaces: availableAcc, // NB_TOT_PLACE_PMR_DISPO
-		TotalStandardSpaces:       totalStd,     // CAP_VEH_NOR
-		TotalAccessibleSpaces:     totalAcc,     // CAP_VEH_PMR
-	}, nil
-}
-
-// ParkingLineConsumer constructs a parking from a slice of strings
-type ParkingLineConsumer struct {
-	parkings map[string]Parking
-}
-
-func makeParkingLineConsumer() *ParkingLineConsumer {
-	return &ParkingLineConsumer{
-		parkings: make(map[string]Parking),
-	}
-}
-
-func (p *ParkingLineConsumer) Consume(line []string, loc *time.Location) error {
-	parking, err := NewParking(line, loc)
-	if err != nil {
-		return err
-	}
-
-	p.parkings[parking.ID] = *parking
-	return nil
-}
-
-func (p *ParkingLineConsumer) Terminate() {}
 
 // Structure and Consumer to creates StopPoint objects based on a line read from a CSV
 type StopPoint struct {
@@ -284,10 +207,6 @@ func NewVehicleOccupancy(rs RouteSchedule, occupancy int) (*VehicleOccupancy, er
 
 // Data manager for all apis
 type DataManager struct {
-	parkings          *map[string]Parking
-	lastParkingUpdate time.Time
-	parkingsMutex     sync.RWMutex
-
 	freeFloatingsContext *freefloatings.FreeFloatingsContext
 
 	stopPoints                   *map[string]StopPoint
@@ -301,21 +220,8 @@ type DataManager struct {
 	equipmentsContext *equipments.EquipmentsContext
 
 	departuresContext *departures.DeparturesContext
-}
 
-func (d *DataManager) UpdateParkings(parkings map[string]Parking) {
-	d.parkingsMutex.Lock()
-	defer d.parkingsMutex.Unlock()
-
-	d.parkings = &parkings
-	d.lastParkingUpdate = time.Now()
-}
-
-func (d *DataManager) GetLastParkingsDataUpdate() time.Time {
-	d.parkingsMutex.RLock()
-	defer d.parkingsMutex.RUnlock()
-
-	return d.lastParkingUpdate
+	parkingsContext *parkings.ParkingsContext
 }
 
 func (d *DataManager) SetEquipmentsContext(equipmentsContext *equipments.EquipmentsContext) {
@@ -342,60 +248,12 @@ func (d *DataManager) GetDeparturesContext() *departures.DeparturesContext {
 	return d.departuresContext
 }
 
-func (d *DataManager) GetParkingsByIds(ids []string) (parkings []Parking, errors []error) {
-	for _, id := range ids {
-		if p, err := d.GetParkingById(id); err == nil {
-			parkings = append(parkings, p)
-		} else {
-			errors = append(errors, err)
-		}
-	}
-	return
+func (d *DataManager) SetParkingsContext(parkingsContext *parkings.ParkingsContext) {
+	d.parkingsContext = parkingsContext
 }
 
-func (d *DataManager) GetParkings() (parkings []Parking, e error) {
-	var mapParkings map[string]Parking
-	{
-		d.parkingsMutex.RLock()
-		defer d.parkingsMutex.RUnlock()
-
-		if d.parkings == nil {
-			e = fmt.Errorf("No parkings in the data")
-			return
-		}
-
-		mapParkings = *d.parkings
-	}
-
-	// Convert Map of parkings to Slice !
-	parkings = make([]Parking, 0, len(mapParkings))
-	for _, p := range mapParkings {
-		parkings = append(parkings, p)
-	}
-
-	return parkings, nil
-}
-
-func (d *DataManager) GetParkingById(id string) (p Parking, e error) {
-	var ok bool
-	{
-		d.parkingsMutex.RLock()
-		defer d.parkingsMutex.RUnlock()
-
-		if d.parkings == nil {
-			e = fmt.Errorf("No parkings in the data")
-			return
-		}
-
-		parkings := *d.parkings
-		p, ok = parkings[id]
-	}
-
-	if !ok {
-		e = fmt.Errorf("No parkings found with id: %s", id)
-	}
-
-	return p, e
+func (d *DataManager) GetParkingsContext() *parkings.ParkingsContext {
+	return d.parkingsContext
 }
 
 func (d *DataManager) ManageVehicleOccupancyStatus(activate bool) {
