+++ type_test.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"sort"
 	"strconv"
 	"testing"
 	"time"
@@ -17,183 +16,6 @@ import (
 	"github.com/CanalTP/forseti/internal/data"
 	"github.com/CanalTP/forseti/internal/equipments"
 )
-
-func TestNewParking(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	location, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-
-	parkingLine := []string{"DECC", "Décines Centre", "2018-09-17 19:29:00", "2018-09-17 19:30:02", "82", "105", "3", "4"}
-
-	p, err := NewParking(parkingLine, location)
-	require.Nil(err)
-	require.NotNil(p)
-
-	assert.Equal("DECC", p.ID)
-	assert.Equal("Décines Centre", p.Label)
-	assert.Equal(time.Date(2018, 9, 17, 19, 29, 0, 0, location), p.UpdatedTime)
-	assert.Equal(82, p.AvailableStandardSpaces)
-	assert.Equal(105, p.TotalStandardSpaces)
-	assert.Equal(3, p.AvailableAccessibleSpaces)
-	assert.Equal(4, p.TotalAccessibleSpaces)
-}
-
-func TestNewParkingWithMissingFields(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	location, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-
-	parkingLine := []string{
-		"DECC", "Décines Centre", "2018-09-17 19:29:00", "2018-09-17 19:30:02", "82", "105"}
-
-	p, err := NewParking(parkingLine, location)
-	assert.Error(err)
-	assert.Nil(p)
-}
-
-func TestNewParkingWithMalformedFields(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	location, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-
-	malformedParkingLines := [][]string{
-		{"DECC", "Décines Centre", "this_should_be_a_date", "2018-09-17 19:30:02", "82", "105", "3", "4"},
-		{"DECC", "Décines Centre", "2018-09-17 19:30:02", "2018-09-17 19:30:02", "this_is_an_int", "105", "3", "4"},
-		{"DECC", "Décines Centre", "2018-09-17 19:30:02", "2018-09-17 19:30:02", "82", "another_int", "3", "4"},
-		{"DECC", "Décines Centre", "2018-09-17 19:30:02", "2018-09-17 19:30:02", "82", "105", "int", "4"},
-		{"DECC", "Décines Centre", "2018-09-17 19:30:02", "2018-09-17 19:30:02", "82", "105", "3", ""},
-	}
-
-	for _, malformedLine := range malformedParkingLines {
-		p, err := NewParking(malformedLine, location)
-		assert.Error(err)
-		assert.Nil(p)
-	}
-}
-
-func TestDataManagerCanGetParkingById(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	loc, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-	updateTime, err := time.ParseInLocation("2006-01-02 15:04:05", "2018-09-17 19:29:00", loc)
-	require.Nil(err)
-
-	var manager DataManager
-	manager.UpdateParkings(map[string]Parking{
-		"toto": {"DECC", "Décines Centre", updateTime, 1, 2, 3, 4},
-	})
-
-	p, err := manager.GetParkingById("toto")
-	assert.Nil(err)
-	assert.Equal("DECC", p.ID)
-	assert.Equal("Décines Centre", p.Label)
-	assert.Equal(time.Date(2018, 9, 17, 19, 29, 0, 0, loc), p.UpdatedTime)
-	assert.Equal(1, p.AvailableStandardSpaces)
-	assert.Equal(2, p.AvailableAccessibleSpaces)
-	assert.Equal(3, p.TotalStandardSpaces)
-	assert.Equal(4, p.TotalAccessibleSpaces)
-}
-
-func TestDataManagerShouldErrorOnEmptyData(t *testing.T) {
-	assert := assert.New(t)
-
-	var manager DataManager
-	p, err := manager.GetParkingById("toto")
-	assert.Error(err)
-	assert.Empty(p)
-}
-
-func TestDataManagerShouldErrorOnEmptyData2(t *testing.T) {
-	assert := assert.New(t)
-
-	var manager DataManager
-	p, errs := manager.GetParkingsByIds([]string{"toto", "titi"})
-	assert.NotEmpty(errs)
-	assert.Empty(p)
-	for _, e := range errs {
-		assert.Error(e)
-	}
-}
-func TestDataManagerCanGetParkingByIds(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	loc, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-	updateTime, err := time.ParseInLocation("2006-01-02 15:04:05", "2018-09-17 19:29:00", loc)
-	require.Nil(err)
-
-	var manager DataManager
-	manager.UpdateParkings(map[string]Parking{
-		"riri":   {"Riri", "First of the name", updateTime, 1, 2, 3, 4},
-		"fifi":   {"Fifi", "Second of the name", updateTime, 1, 2, 3, 4},
-		"loulou": {"Loulou", "Third of the name", updateTime, 1, 2, 3, 4},
-	})
-
-	p, errs := manager.GetParkingsByIds([]string{"riri", "loulou"})
-	assert.Nil(errs)
-	require.Len(p, 2)
-	assert.Equal("Riri", p[0].ID)
-	assert.Equal("Loulou", p[1].ID)
-}
-
-func TestDataManagerCanPartiallyGetParkingByIds(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	loc, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-	updateTime, err := time.ParseInLocation("2006-01-02 15:04:05", "2018-09-17 19:29:00", loc)
-	require.Nil(err)
-
-	var manager DataManager
-	manager.UpdateParkings(map[string]Parking{
-		"riri":   {"Riri", "First of the name", updateTime, 1, 2, 3, 4},
-		"fifi":   {"Fifi", "Second of the name", updateTime, 1, 2, 3, 4},
-		"loulou": {"Loulou", "Third of the name", updateTime, 1, 2, 3, 4},
-	})
-
-	p, errs := manager.GetParkingsByIds([]string{"fifi", "donald"})
-	require.Len(p, 1)
-	assert.Equal("Fifi", p[0].ID)
-
-	require.Len(errs, 1)
-	assert.Error(errs[0])
-}
-
-func TestDataManagerGetParking(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	loc, err := time.LoadLocation("Europe/Paris")
-	require.Nil(err)
-	updateTime, err := time.ParseInLocation("2006-01-02 15:04:05", "2018-09-17 19:29:00", loc)
-	require.Nil(err)
-
-	var manager DataManager
-	manager.UpdateParkings(map[string]Parking{
-		"riri":   {"Riri", "First of the name", updateTime, 1, 2, 3, 4},
-		"fifi":   {"Fifi", "Second of the name", updateTime, 1, 2, 3, 4},
-		"loulou": {"Loulou", "Third of the name", updateTime, 1, 2, 3, 4},
-	})
-
-	parkings, err := manager.GetParkings()
-	require.Nil(err)
-	require.Len(parkings, 3)
-
-	sort.Sort(ByParkingId(parkings))
-	assert.Equal("Fifi", parkings[0].ID)
-	assert.Equal("Loulou", parkings[1].ID)
-	assert.Equal("Riri", parkings[2].ID)
-}
 
 func TestData(t *testing.T) {
 	assert := assert.New(t)
